use 1 file for azure batches handling
